Início das validações do frontend

+++ routes/web.php
@@ -12,54 +12,50 @@ Route::get('/', function () {
     return view('/sistema/home');
 });
 
-
-
 /*
     Vendas
 */
-Route::get('/vendas', [VendaController::class, 'index']);
-Route::get('/vendas/detalhes/{id?}', [VendaController::class, 'show']);
-Route::get('/cadastro/vendas', [VendaController::class, 'create']);
-Route::post('/vendas/store', [VendaController::class, 'store']);
-Route::get('/editar/vendas/{id}', [VendaController::class, 'edit']);
-Route::post('/vendas/update/{id}', [VendaController::class, 'update']);
-Route::get('/deletar/vendas/{id}', [VendaController::class, 'delete']);
-Route::delete('/vendas/destroy/{id}', [VendaController::class, 'destroy']);
-
-
+Route::get('/vendas', [VendaController::class, 'index'])->middleware('auth');
+Route::get('/vendas/detalhes/{id?}', [VendaController::class, 'show'])->middleware('auth');
+Route::get('/cadastro/vendas', [VendaController::class, 'create'])->middleware('auth');
+Route::post('/vendas/store', [VendaController::class, 'store'])->middleware('auth');
+Route::get('/editar/vendas/{id}', [VendaController::class, 'edit'])->middleware('auth');
+Route::post('/vendas/update/{id}', [VendaController::class, 'update'])->middleware('auth');
+Route::get('/deletar/vendas/{id}', [VendaController::class, 'delete'])->middleware('auth');
+Route::delete('/vendas/destroy/{id}', [VendaController::class, 'destroy'])->middleware('auth');
 
 /*
     Clientes
 */
-Route::get('/clientes', [ClienteController::class, 'index']);
-Route::get('/cadastro/clientes', [ClienteController::class, 'create']);
-Route::post('/clientes/store', [ClienteController::class, 'store']);
-Route::get('/editar/clientes/{id}', [ClienteController::class, 'edit']);
-Route::post('/clientes/update/{id}', [ClienteController::class, 'update']);
-Route::get('/deletar/clientes/{id}', [ClienteController::class, 'delete']);
-Route::delete('/clientes/destroy/{id}', [ClienteController::class, 'destroy']);
+Route::get('/clientes', [ClienteController::class, 'index'])->middleware('auth');
+Route::get('/cadastro/clientes', [ClienteController::class, 'create'])->middleware('auth');
+Route::post('/clientes/store', [ClienteController::class, 'store'])->middleware('auth');
+Route::get('/editar/clientes/{id}', [ClienteController::class, 'edit'])->middleware('auth');
+Route::post('/clientes/update/{id}', [ClienteController::class, 'update'])->middleware('auth');
+Route::get('/deletar/clientes/{id}', [ClienteController::class, 'delete'])->middleware('auth');
+Route::delete('/clientes/destroy/{id}', [ClienteController::class, 'destroy'])->middleware('auth');
 
 /*
     Produtos
 */
-Route::get('/produtos', [ProdutoController::class, 'index']);
-Route::get('/cadastro/produtos', [ProdutoController::class, 'create']);
-Route::post('/produtos/store', [ProdutoController::class, 'store']);
-Route::get('/editar/produtos/{id}', [ProdutoController::class, 'edit']);
-Route::post('/produtos/update/{id}', [ProdutoController::class, 'update']);
-Route::get('/deletar/produtos/{id}', [ProdutoController::class, 'delete']);
-Route::delete('/produtos/destroy/{id}', [ProdutoController::class, 'destroy']);
+Route::get('/produtos', [ProdutoController::class, 'index'])->middleware('auth');
+Route::get('/cadastro/produtos', [ProdutoController::class, 'create'])->middleware('auth');
+Route::post('/produtos/store', [ProdutoController::class, 'store'])->middleware('auth');
+Route::get('/editar/produtos/{id}', [ProdutoController::class, 'edit'])->middleware('auth');
+Route::post('/produtos/update/{id}', [ProdutoController::class, 'update'])->middleware('auth');
+Route::get('/deletar/produtos/{id}', [ProdutoController::class, 'delete'])->middleware('auth');
+Route::delete('/produtos/destroy/{id}', [ProdutoController::class, 'destroy'])->middleware('auth');
 
 /*
     Vendedores
 */
-Route::get('/vendedores', [UserController::class, 'index']);
-Route::get('/cadastro/vendedores', [UserController::class, 'create']);
-Route::post('/vendedores/store', [UserController::class, 'store']);
-Route::get('/editar/vendedores/{id}', [UserController::class, 'edit']);
-Route::post('/vendedores/update/{id}', [UserController::class, 'update']);
-Route::get('/deletar/vendedores/{id}', [UserController::class, 'delete']);
-Route::delete('/vendedores/destroy/{id}', [UserController::class, 'destroy']);
+Route::get('/vendedores', [UserController::class, 'index'])->middleware('auth');
+Route::get('/cadastro/vendedores', [UserController::class, 'create'])->middleware('auth');
+Route::post('/vendedores/store', [UserController::class, 'store'])->middleware('auth');
+Route::get('/editar/vendedores/{id}', [UserController::class, 'edit'])->middleware('auth');
+Route::post('/vendedores/update/{id}', [UserController::class, 'update'])->middleware('auth');
+Route::get('/deletar/vendedores/{id}', [UserController::class, 'delete'])->middleware('auth');
+Route::delete('/vendedores/destroy/{id}', [UserController::class, 'destroy'])->middleware('auth');
 
 /*
     Login
